Migrated to Spring JPA

+++ checksystem-core/src/main/java/ru/clevertec/checksystem/core/common/builder/discount/checkitem/IThresholdPercentageCheckItemDiscountBuilder.java
@@ -3,5 +3,5 @@ package ru.clevertec.checksystem.core.common.builder.discount.checkitem;
 public interface IThresholdPercentageCheckItemDiscountBuilder
         extends IPercentageCheckItemDiscountBuilder<IThresholdPercentageCheckItemDiscountBuilder> {
 
-    IThresholdPercentageCheckItemDiscountBuilder setThreshold(int threshold);
+    IThresholdPercentageCheckItemDiscountBuilder setThreshold(Long threshold);
 }
